@@ -22,6 +22,6 @@
 		"@types/jest": "^24.0.15",
 		"jest": "^24.8.0",
 		"ts-jest": "^24.0.2",
-		"typescript": "^3.5.2"
+		"typescript": "^3.7.0-beta"
 	}
 }
