@@ -1,2 +1,4 @@
 require("./polyfills/index.js");
-require("./cli/index.js")
+
+import cogniProgram from "./cli";
+cogniProgram.takeArgv(process.argv.slice(1));
